requirements.txt
# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.3
scipy>=1.10.1

# Visualization
matplotlib>=3.7.1
seaborn>=0.12.2

# Machine Learning
scikit-learn>=1.3.0

# API and Web
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0

# Data Collection
praw>=7.7.1
yfinance>=0.2.36

# Other utilities
joblib>=1.3.0
textblob>=0.17.1

PyGithub==1.59.1
pandas_datareader==0.10.0
fredapi==0.5.1
tweepy==4.14.0
schedule==1.1.0

flask>=2.0.1
flask-sqlalchemy>=2.5.1
flask-limiter>=2.4.0
gunicorn>=20.1.0
prometheus-client>=0.11.0  # For monitoring

# Additional requirements
flask-jwt-extended>=4.4.4
flask-migrate>=4.0.4
psycopg2-binary>=2.9.6
flask-cors>=3.0.10 